zml/tests: ensure all Zig tests are running (#14)
This PR restore skipped tests. 
It was skipped because we managed to have a CI infrastructure for the
public release. There was a problem regarding computation precision 
on ubuntu/cpu platform so it was skipped. It's now fixed using `f32` 
instead of `f16` precision so we have lower diffs.
I also added `std.testing.refAllDecls` on each main namespace to ensure
we run every declared tests in `zml/`.
We have to rework about how we want to reference all tests but for now
it does the job.