add DotAlgorim support to dot_general
for now it's not exposed to users because the API is a bit limited.
in particular only a subset of algorithm are supported by openxla,
and only a subset of that is supported by chips.

I was able to successfully change the accumulation type for dot on Cuda,
but want to experiment more before deciding what to expose.