zml: makeBlocks returns the inner tensor to propagate tags (#92)
Now zml.ops.makeBlock returns both the mlir.Block created and also the
Tensors returned by `func`.

The returned tensors should not be returned to the user,
because their `mlir.Value` must not escape the block that created them.
But their shapes/tags can be safely propagated further.

This makes `zml.ops.makeBlock` a bit more footgunny, but also it's a
utility function that most users shouldn't use. And it allows to use
tags created inside the block and apply them to the return value. In
particular tags and sharding info should now be propagated more
consistently for while loops, summation, reduceWindow, ...

I've also enabled tests to run on non-cpu platform.
this is useful to figure out if a specific platform has limitation in
it's pjrt support