runtimes/cuda: rework sandboxing and dynamic symbol renaming
This change reworks the sandboxing and dynamic symbol renaming for the cuda runtime.

Switch to the jax-cuda-pjrt pre-built PJRT plugin until we fix the pjrt artifacts build.
Most notably because our build of the plugin didn't include the right compute capabilities.

Take the opportunity to bump the `cuda=12.6.2` `cudnn=9.5.1`.

Fixes #19