runtimes/rocm: sandbox ROCm dependencies (#276)
ROCm counterpart of #275 

The most important thing with this change is that ROCm dependencies
(along with the PJRT plugin) must be loaded in the main thread.

This is because some ROCm dependencies (amdhip64) use Thread Local
Storage in the destructor of static C++ instances.
Because all dependencies are unloaded at exit of the main program,
destructors are run on whatever thread exited (in general the main
thread), and any access to the TLS will except values from a different
threads....